fix(directory): remove domain main folder
