+++ src/main/java/bungae/thunder/cakey/cake/domain/Cake.java
@@ -1,0 +1,12 @@
+package bungae.thunder.cakey.cake.domain;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class Cake {
+    private Long id;
+    private Integer year;
+    private Long userId;
+}
