+++ src/main/java/bungae/thunder/cakey/cake/repository/CakeRepository.java
@@ -1,0 +1,20 @@
+package bungae.thunder.cakey.cake.repository;
+
+import bungae.thunder.cakey.cake.domain.Cake;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface CakeRepository {
+    Cake save(Cake cake);
+
+    Optional<Cake> findById(Long id);
+
+    Optional<Cake> findOneByUserId(Long userId);
+
+    List<Cake> findAllByUserId(Long userId);
+
+    Optional<Cake> findByUserIdAndYear(Long userId, Integer year);
+
+    List<Cake> findAll();
+}
