--- src/main/java/bungae/thunder/cakey/domain/cake/domain/Cake.java
@@ -1,12 +1,0 @@
-package bungae.thunder.cakey.domain.cake.domain;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class Cake {
-    private Long id;
-    private Integer year;
-    private Long userId;
-}
