--- src/main/java/bungae/thunder/cakey/domain/report/repository/ReportRepository.java
@@ -1,19 +1,0 @@
-package bungae.thunder.cakey.domain.report.repository;
-
-import bungae.thunder.cakey.domain.report.domain.Report;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface ReportRepository {
-    Report save(Report report);
-
-    Optional<Report> findById(Long id);
-
-    List<Report> findAllByMessageId(Long messageId);
-
-    List<Report> findAll();
-
-}
