--- src/main/java/bungae/thunder/cakey/domain/user/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package bungae.thunder.cakey.domain.user.repository;
-
-import bungae.thunder.cakey.domain.user.domain.User;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface UserRepository {
-    User save(User user);
-
-    Optional<User> findById(Long id);
-
-    Optional<User> findByName(String name);
-
-    List<User> findAll();
-}
