+++ src/main/java/bungae/thunder/cakey/message/domain/Message.java
@@ -1,0 +1,15 @@
+package bungae.thunder.cakey.message.domain;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class Message {
+    private Long id;
+    private String contents;
+    private String reply;
+    private String audioUrl;
+    private Long cakeId;
+    private Long senderId;
+}
