+++ src/main/java/bungae/thunder/cakey/report/repository/ReportRepository.java
@@ -1,0 +1,19 @@
+package bungae.thunder.cakey.report.repository;
+
+import bungae.thunder.cakey.report.domain.Report;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.Optional;
+
+@Repository
+public interface ReportRepository {
+    Report save(Report report);
+
+    Optional<Report> findById(Long id);
+
+    List<Report> findAllByMessageId(Long messageId);
+
+    List<Report> findAll();
+
+}
