+++ src/main/java/bungae/thunder/cakey/user/repository/UserRepository.java
@@ -1,0 +1,16 @@
+package bungae.thunder.cakey.user.repository;
+
+import bungae.thunder.cakey.user.domain.User;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface UserRepository {
+    User save(User user);
+
+    Optional<User> findById(Long id);
+
+    Optional<User> findByName(String name);
+
+    List<User> findAll();
+}
